check code

+++ src/services/contacts.js
@@ -8,11 +8,12 @@ export const allContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find({ userId });
 
   if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
@@ -38,18 +39,20 @@ export const allContacts = async ({
   };
 };
 
-export const oneContactForId = async (id) => {
-  return await ContactsCollection.findById(id);
+export const oneContactForId = async ({ contactId, userId }) => {
+  return await ContactsCollection.findOne({ _id: contactId, userId });
 };
 
 export const createContact = async (contact) => {
   return await ContactsCollection.create(contact);
 };
 
-export const updateContactForId = async (_id, contact) => {
-  return ContactsCollection.findOneAndUpdate({ _id }, contact, { new: true });
+export const updateContactForId = async (_id, contact, userId) => {
+  return ContactsCollection.findOneAndUpdate({ _id, userId }, contact, {
+    new: true,
+  });
 };
 
-export const deleteContactForId = async (_id) => {
-  return ContactsCollection.findOneAndDelete({ _id });
+export const deleteContactForId = async (_id, userId) => {
+  return ContactsCollection.findOneAndDelete({ _id, userId });
 };
